requirements.txt
flask
langchain
langchain-google-genai
requests
pydantic
python-dotenv
